updateproject

+++ springcloud2020/cloud-provider-payment8001/src/main/java/com/lk/springcloud/dao/PaymentDao.java
@@ -12,5 +12,5 @@ import org.apache.ibatis.annotations.Param;
 public interface PaymentDao {
     public int create(Payment payment);
 
-    public Payment getPaymentById(@Param("id") Integer id);
+    public Payment getPaymentById(@Param("id") Long id);
 }
